# api/docs/swagger.yaml
basePath: /
definitions:
  proto.BranchesRes:
    properties:
      contact:
        type: string
      created_at:
        type: string
      description:
        type: string
      google_url:
        type: string
      id:
        type: string
      name:
        type: string
      yandex_url:
        type: string
    type: object
  proto.CertificateRes:
    properties:
      cefr_level:
        type: string
      certificate_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      ielts_score:
        type: number
      name:
        type: string
    type: object
...
  proto.CreateBranches:
    properties:
      contact:
        type: string
      description:
        type: string
      google_url:
        type: string
      name:
        type: string
      yandex_url:
        type: string
    type: object
  proto.CreateCertificate:
    properties:
      cefr_level:
        type: string
      certificate_url:
        type: string
      description:
        type: string
      ielts_score:
        type: number
      name:
        type: string
    type: object
...
  proto.GetListBranchesRes:
    properties:
      branches:
        items:
          $ref: '#/definitions/proto.BranchesRes'
        type: array
      total_count:
        type: integer
    type: object
  proto.GetListCertificateRes:
    properties:
      certificates:
        items:
          $ref: '#/definitions/proto.CertificateRes'
        type: array
      total_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Pima
  title: Pima
  version: "1.0"
paths:
  /branches/create:
    post:
      consumes:
      - application/json
      description: Create branches
      parameters:
      - description: Create branches
        in: body
        name: branches
        required: true
        schema:
          $ref: '#/definitions/proto.CreateBranches'
      produces:
      - application/json
      responses:
        "200":
          description: Create branches successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create branches
      tags:
      - branches
  /branches/delete:
    delete:
      consumes:
      - application/json
      description: Delete a branches by ID
      parameters:
      - description: Branches ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Branches deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a branches
      tags:
      - branches
  /branches/get:
    get:
      consumes:
      - application/json
      description: Get an branches by their ID
      parameters:
      - description: Branches ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.BranchesRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get branches by ID
      tags:
      - branches
  /branches/list:
    get:
      consumes:
      - application/json
      description: Get all branches with optional filtering
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.GetListBranchesRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all branches
      tags:
      - branches
  /branches/update:
    put:
      consumes:
      - application/json
      description: Update branches
      parameters:
      - description: Branches ID
        in: query
        name: id
        type: string
      - description: Branches Update Details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/proto.CreateBranches'
      produces:
      - application/json
      responses:
        "200":
          description: Branches updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update branches
      tags:
      - branches
  /certificates/create:
    post:
      consumes:
      - application/json
      description: Create a new certificate
      parameters:
      - description: Create certificate
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/proto.CreateCertificate'
      produces:
      - application/json
      responses:
        "200":
          description: Certificate created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a certificate
      tags:
      - certificates
  /certificates/delete:
    delete:
      consumes:
      - application/json
      description: Delete a certificate by ID
      parameters:
      - description: Certificate ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Certificate deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a certificate
      tags:
      - certificates
  /certificates/get:
    get:
      consumes:
      - application/json
      description: Get a certificate by its ID
      parameters:
      - description: Certificate ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.CertificateRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get certificate by ID
      tags:
      - certificates
  /certificates/list:
    get:
      consumes:
      - application/json
      description: Fetch a list of certificates with optional filters for name, IELTS
        score, CEFR level, and pagination parameters.
      parameters:
      - description: Filter by certificate name
        in: query
        name: name
        type: string
      - description: Filter by IELTS score
        in: query
        name: ielts_score
        type: number
      - description: Filter by CEFR level
        in: query
        name: cefr_level
        type: string
      - description: Number of records to retrieve (default is 10)
        in: query
        name: limit
        type: integer
      - description: Number of records to skip for pagination (default is 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A list of certificates matching the filters
          schema:
            $ref: '#/definitions/proto.GetListCertificateRes'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Retrieve a list of certificates
      tags:
      - certificates
  /certificates/update:
    put:
      consumes:
      - application/json
      description: Update certificate details
      parameters:
      - description: Certificate ID
        in: query
        name: id
        required: true
        type: string
      - description: Update certificate
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/proto.CreateCertificate'
      produces:
      - application/json
      responses:
        "200":
          description: Certificate updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a certificate
      tags:
      - certificates
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"